Changed endpoint method should return different HttpStatus

+++ src/main/kotlin/kr/co/redbrush/microservice/app/controller/AccountController.kt
@@ -3,8 +3,9 @@ package kr.co.redbrush.microservice.app.controller
 import kr.co.redbrush.microservice.app.data.Account
 import kr.co.redbrush.microservice.app.service.AccountService
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import java.util.concurrent.ConcurrentHashMap
 
 @RestController
 class AccountController {
@@ -12,7 +13,12 @@ class AccountController {
     private lateinit var accountService: AccountService
 
     @GetMapping("/account/{id}")
-    fun getAccount(@PathVariable id: String) = accountService.getAccount(id)
+    fun getAccount(@PathVariable id: String) : ResponseEntity<Account?> {
+        val account = accountService.getAccount(id)
+        val status = if (account == null) HttpStatus.NOT_FOUND else HttpStatus.OK
+
+        return ResponseEntity(account, status)
+    }
 
     @GetMapping("/accounts")
     fun getAccounts(@RequestParam(required = false, defaultValue = "") idFilter: String) : List<Account> {
@@ -20,17 +26,33 @@ class AccountController {
     }
 
     @PostMapping("/account")
-    fun createAccount(@RequestBody account : Account) {
+    fun createAccount(@RequestBody account : Account): ResponseEntity<Unit> {
         accountService.createAccount(account)
+
+        return ResponseEntity(Unit, HttpStatus.CREATED)
     }
 
     @PutMapping("/account/{id}")
-    fun updateAccount(@PathVariable id: String, @RequestBody account: Account) {
-        accountService.updateAccount(id, account)
+    fun updateAccount(@PathVariable id: String, @RequestBody account: Account): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+
+        if (accountService.getAccount(id) != null) {
+            accountService.updateAccount(id, account)
+            status = HttpStatus.ACCEPTED
+        }
+
+        return ResponseEntity(Unit, status)
     }
 
     @DeleteMapping("/account/{id}")
-    fun deleteAccount(@PathVariable id: String) {
-        accountService.deleteAccount(id)
+    fun deleteAccount(@PathVariable id: String): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+
+        if (accountService.getAccount(id) != null) {
+            accountService.deleteAccount(id)
+            status = HttpStatus.OK
+        }
+
+        return ResponseEntity(Unit, status)
     }
 }
